+++ pystac_client/item_search.py
@@ -140,12 +140,12 @@ class ItemSearch:
             the response, it will automatically retry with
             ``"GET"`` for all subsequent requests.
         max_items : The maximum number of items to return from the search, even
-            if there are more matching results. This client to limit the
+            if there are more matching results. This allows the client to limit the
             total number of Items returned from the :meth:`items`,
             :meth:`item_collections`, and :meth:`items_as_dicts methods`. The client
             will continue to request pages of items until the number of max items is
-            reached. This parameter defaults to 100. Setting this to ``None`` will
-            allow iteration over a possibly very large number of results.
+            reached. By default (``max_items=None``) all items matching the query
+            will be returned.
         stac_io: An instance of StacIO for retrieving results. Normally comes
             from the Client that returns this ItemSearch client: An instance of a
             root Client used to set the root on resulting Items.
@@ -217,7 +217,7 @@ class ItemSearch:
         url: str,
         *,
         method: Optional[str] = "POST",
-        max_items: Optional[int] = DEFAULT_LIMIT_AND_MAX_ITEMS,
+        max_items: Optional[int] = None,
         stac_io: Optional[StacApiIO] = None,
         client: Optional["_client.Client"] = None,
         limit: Optional[int] = DEFAULT_LIMIT_AND_MAX_ITEMS,